// packages/app-backend-grpc/src/proto/google/protobuf/wrappers.ts
export interface StringValue {
  /**
   * The string value.
   *
   * @generated from protobuf field: string value = 1;
   */
  value: string;
}
// ...
class StringValue$Type extends MessageType<StringValue> {
  constructor() {
    super("google.protobuf.StringValue", [
      { no: 1, name: "value", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
    ]);
  }
  /**
   * Encode `StringValue` to JSON string.
   */
  internalJsonWrite(
    message: StringValue,
    options: JsonWriteOptions
  ): JsonValue {
    return message.value;
  }
  /**
   * Decode `StringValue` from JSON string.
   */
  internalJsonRead(
    json: JsonValue,
    options: JsonReadOptions,
    target?: StringValue
  ): StringValue {
    if (!target) target = this.create();
    target.value = this.refJsonReader.scalar(
      json,
      9,
      undefined,
      "value"
    ) as string;
    return target;
  }
}
/**
 * @generated MessageType for protobuf message google.protobuf.StringValue
 */
export const StringValue = new StringValue$Type();